// src/app/contactmanager/models/product.model.ts
export class Product {
    weatherAlertID : string;
    severity: string;
    disasterType: string;
    disasterOrigin: string;
    impactedArea : string;
    imdDateTime : Date;
    imdComment:string;
    stateDateTime: Date;
    stateComment: string;
    stateStatus: string;
    centralDateTime: Date;
    centralComment:string;
    centralStatus:string;
}